refactor(signal): split config resolver tests
Keep the Signal config resolver small after hardening by moving resolver helpers and tests into focused submodules. This addresses merge-readiness quality-audit structural findings without changing behavior.


Copilot-Session: 29147940-80a3-4c99-9db5-8b796a13cfcd
