+++ crates/amplihack-signal/src/config.rs
@@ -8,6 +8,8 @@
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
+mod resolver;
+
 /// Environment variable names (single source of truth for env + tests).
 pub const ENV_ENDPOINT: &str = "AMPLIHACK_SIGNAL_ENDPOINT";
 pub const ENV_ACCOUNT: &str = "AMPLIHACK_SIGNAL_ACCOUNT";
@@ -100,115 +102,19 @@ impl SignalConfig {
         };
         let toml_table = toml_val.as_ref().and_then(toml::Value::as_table);
 
-        // env > TOML for a string-valued setting.
-        let get_str = |env_key: &str, toml_key: &str| -> Option<String> {
-            if let Some(v) = env.get(env_key) {
-                return Some(v.clone());
-            }
-            toml_table
-                .and_then(|t| t.get(toml_key))
-                .and_then(toml::Value::as_str)
-                .map(str::to_string)
-        };
-        let is_present = |env_key: &str, toml_key: &str| -> bool {
-            env.contains_key(env_key) || toml_table.is_some_and(|t| t.contains_key(toml_key))
-        };
+        let endpoint = resolver::get_str(env, toml_table, ENV_ENDPOINT, "endpoint")
+            .ok_or(ConfigError::MissingRequired("endpoint"))?;
+        resolver::validate_endpoint(&endpoint)?;
 
-        let endpoint =
-            get_str(ENV_ENDPOINT, "endpoint").ok_or(ConfigError::MissingRequired("endpoint"))?;
-        validate_endpoint(&endpoint)?;
+        let account = resolver::get_str(env, toml_table, ENV_ACCOUNT, "account")
+            .ok_or(ConfigError::MissingRequired("account"))?;
+        resolver::validate_e164(&account)?;
 
-        let account =
-            get_str(ENV_ACCOUNT, "account").ok_or(ConfigError::MissingRequired("account"))?;
-        validate_e164(&account)?;
-
-        // The allowlist key MUST be present (absence is a hard error; an
-        // explicitly-empty allowlist is a valid, deliberate fail-closed config).
-        if !is_present(ENV_ALLOWLIST, "allowlist") {
-            return Err(ConfigError::MissingRequired("allowlist"));
-        }
-        let allowlist: Vec<String> = if let Some(csv) = env.get(ENV_ALLOWLIST) {
-            parse_allowlist_csv(csv)?
-        } else {
-            let mut out = Vec::new();
-            let Some(raw_allowlist) = toml_table.and_then(|t| t.get("allowlist")) else {
-                return Err(ConfigError::MissingRequired("allowlist"));
-            };
-            let arr = raw_allowlist.as_array().ok_or_else(|| {
-                ConfigError::Toml(format!(
-                    "invalid allowlist: expected array, got {raw_allowlist}"
-                ))
-            })?;
-            for item in arr {
-                let s = item.as_str().ok_or_else(|| {
-                    ConfigError::Toml(format!(
-                        "invalid allowlist entry: expected string, got {item}"
-                    ))
-                })?;
-                let s = s.trim();
-                if s.is_empty() {
-                    continue;
-                }
-                validate_e164(s)?;
-                out.push(s.to_string());
-            }
-            out
-        };
-
-        let own_device_id = match env.get(ENV_OWN_DEVICE_ID) {
-            Some(v) => Some(
-                v.trim()
-                    .parse::<u32>()
-                    .map_err(|_| ConfigError::InvalidNumber {
-                        key: ENV_OWN_DEVICE_ID,
-                        value: v.clone(),
-                    })?,
-            ),
-            None => match toml_table.and_then(|t| t.get("own_device_id")) {
-                Some(v) => {
-                    let i = v.as_integer().ok_or_else(|| ConfigError::InvalidNumber {
-                        key: ENV_OWN_DEVICE_ID,
-                        value: v.to_string(),
-                    })?;
-                    Some(u32::try_from(i).map_err(|_| ConfigError::InvalidNumber {
-                        key: ENV_OWN_DEVICE_ID,
-                        value: i.to_string(),
-                    })?)
-                }
-                None => None,
-            },
-        };
-
-        // signal-cli's own linked-device id, when configured, must be a real
-        // linked device (`>= 2`). Device `1` is the operator's primary phone and
-        // must never be treated as the bot's own echo source.
-        match own_device_id {
-            Some(d) if d < 2 => {
-                return Err(ConfigError::InvalidNumber {
-                    key: ENV_OWN_DEVICE_ID,
-                    value: d.to_string(),
-                });
-            }
-            _ => {}
-        }
-
-        let reuse_rolling_group = match env.get(ENV_REUSE_ROLLING_GROUP) {
-            Some(v) => parse_bool_env(ENV_REUSE_ROLLING_GROUP, v)?,
-            None => match toml_table.and_then(|t| t.get("reuse_rolling_group")) {
-                Some(v) => v.as_bool().ok_or_else(|| {
-                    ConfigError::Toml(format!("invalid boolean setting reuse_rolling_group: {v}"))
-                })?,
-                None => false,
-            },
-        };
-
-        let rolling_group_id = get_str(ENV_ROLLING_GROUP_ID, "rolling_group_id")
-            .map(|v| v.trim().to_string())
-            .filter(|v| !v.is_empty());
-        if reuse_rolling_group && rolling_group_id.is_none() {
-            return Err(ConfigError::MissingRequired("rolling_group_id"));
-        }
-
+        let allowlist = resolver::resolve_allowlist(env, toml_table)?;
+        let own_device_id = resolver::resolve_own_device_id(env, toml_table)?;
+        let reuse_rolling_group = resolver::resolve_reuse_rolling_group(env, toml_table)?;
+        let rolling_group_id =
+            resolver::resolve_rolling_group_id(env, toml_table, reuse_rolling_group)?;
         Ok(SignalConfig {
             endpoint,
             account,
@@ -269,435 +175,5 @@ pub fn resolve_config_source(
     Ok(None)
 }
 
-/// Parse a comma-separated allowlist, trimming and dropping empty entries and
-/// validating every surviving entry as E.164.
-fn parse_allowlist_csv(csv: &str) -> Result<Vec<String>, ConfigError> {
-    let mut out = Vec::new();
-    for part in csv.split(',') {
-        let part = part.trim();
-        if part.is_empty() {
-            continue;
-        }
-        validate_e164(part)?;
-        out.push(part.to_string());
-    }
-    Ok(out)
-}
-
-/// Validate an E.164 phone number: `+` followed by 1..=15 ASCII digits.
-fn validate_e164(s: &str) -> Result<(), ConfigError> {
-    let ok = s.starts_with('+') && {
-        let digits = &s[1..];
-        !digits.is_empty() && digits.len() <= 15 && digits.bytes().all(|b| b.is_ascii_digit())
-    };
-    if ok {
-        Ok(())
-    } else {
-        Err(ConfigError::InvalidE164(s.to_string()))
-    }
-}
-
-/// Validate a `host:port` endpoint: non-empty host and non-zero `u16` port.
-/// IPv6 literals must use the standard bracket form (`[::1]:7583`) so the host
-/// boundary is unambiguous.
-fn validate_endpoint(s: &str) -> Result<(), ConfigError> {
-    let (host, port) = if let Some(rest) = s.strip_prefix('[') {
-        let (host, rest) = rest
-            .split_once("]:")
-            .ok_or_else(|| ConfigError::InvalidEndpoint(s.to_string()))?;
-        (host, rest)
-    } else if let Some((host, port)) = s.rsplit_once(':') {
-        if host.contains(':') {
-            return Err(ConfigError::InvalidEndpoint(s.to_string()));
-        }
-        (host, port)
-    } else {
-        return Err(ConfigError::InvalidEndpoint(s.to_string()));
-    };
-    if !host.is_empty() && port.parse::<u16>().is_ok_and(|p| p != 0) {
-        return Ok(());
-    }
-    Err(ConfigError::InvalidEndpoint(s.to_string()))
-}
-
-/// Parse common boolean env tokens. Empty is retained as a safe explicit false
-/// for the isolation default; unknown non-empty tokens are configuration errors.
-fn parse_bool_env(key: &'static str, v: &str) -> Result<bool, ConfigError> {
-    match v.trim().to_ascii_lowercase().as_str() {
-        "1" | "true" | "yes" | "on" => Ok(true),
-        "0" | "false" | "no" | "off" | "" => Ok(false),
-        _ => Err(ConfigError::Toml(format!(
-            "invalid boolean setting {key}: {v}"
-        ))),
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn env(pairs: &[(&str, &str)]) -> HashMap<String, String> {
-        pairs
-            .iter()
-            .map(|(k, v)| ((*k).to_string(), (*v).to_string()))
-            .collect()
-    }
-
-    #[test]
-    fn env_only_minimal_valid_config() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001,+15551230002"),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, None).expect("valid config");
-        assert_eq!(cfg.endpoint, "127.0.0.1:7583");
-        assert_eq!(cfg.account, "+15551230000");
-        assert_eq!(cfg.allowlist, vec!["+15551230001", "+15551230002"]);
-        assert_eq!(cfg.own_device_id, None);
-        assert!(!cfg.reuse_rolling_group);
-        assert_eq!(cfg.rolling_group_id, None);
-    }
-
-    #[test]
-    fn own_device_id_parsed_from_env() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-            (ENV_OWN_DEVICE_ID, "3"),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, None).unwrap();
-        assert_eq!(cfg.own_device_id, Some(3));
-    }
-
-    #[test]
-    fn own_device_id_below_two_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-            (ENV_OWN_DEVICE_ID, "1"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(
-            matches!(err, ConfigError::InvalidNumber { key, .. } if key == ENV_OWN_DEVICE_ID),
-            "expected InvalidNumber for own_device_id, got {err:?}"
-        );
-    }
-
-    #[test]
-    fn toml_own_device_id_below_two_is_error() {
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001"]
-            own_device_id = -1
-        "#;
-        let err = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap_err();
-        assert!(
-            matches!(err, ConfigError::InvalidNumber { key, .. } if key == ENV_OWN_DEVICE_ID),
-            "expected InvalidNumber for own_device_id, got {err:?}"
-        );
-    }
-
-    #[test]
-    fn missing_endpoint_is_error() {
-        let e = env(&[
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::MissingRequired("endpoint")));
-    }
-
-    #[test]
-    fn missing_account_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::MissingRequired("account")));
-    }
-
-    #[test]
-    fn absent_allowlist_is_error_no_silent_default() {
-        // The allowlist key is required to be *present*. Absence must error —
-        // never silently default to "allow everyone".
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::MissingRequired("allowlist")));
-    }
-
-    #[test]
-    fn present_but_empty_allowlist_is_valid_fail_closed() {
-        // Explicitly empty is a *valid* deliberate config: "accept no inbound".
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, ""),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, None).expect("empty allowlist is valid");
-        assert!(cfg.allowlist.is_empty());
-    }
-
-    #[test]
-    fn invalid_e164_account_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "5551230000"), // missing '+'
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::InvalidE164(_)));
-    }
-
-    #[test]
-    fn invalid_e164_in_allowlist_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001,not-a-number"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::InvalidE164(_)));
-    }
-
-    #[test]
-    fn invalid_endpoint_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1"), // no port
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::InvalidEndpoint(_)));
-    }
-
-    #[test]
-    fn endpoint_rejects_unbracketed_ipv6_and_port_zero() {
-        for endpoint in ["::1", "fe80::1", "127.0.0.1:0"] {
-            let e = env(&[
-                (ENV_ENDPOINT, endpoint),
-                (ENV_ACCOUNT, "+15551230000"),
-                (ENV_ALLOWLIST, "+15551230001"),
-            ]);
-            let err = SignalConfig::from_sources(&e, None).unwrap_err();
-            assert!(
-                matches!(err, ConfigError::InvalidEndpoint(_)),
-                "{endpoint} should be rejected, got {err:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn endpoint_accepts_bracketed_ipv6() {
-        let e = env(&[
-            (ENV_ENDPOINT, "[::1]:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, None).expect("bracketed IPv6 endpoint is valid");
-        assert_eq!(cfg.endpoint, "[::1]:7583");
-    }
-
-    #[test]
-    fn toml_supplies_values_when_env_absent() {
-        let toml = r#"
-            endpoint = "10.0.0.5:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001", "+15551230002"]
-            own_device_id = 2
-            reuse_rolling_group = true
-            rolling_group_id = "grp-rolling=="
-        "#;
-        let cfg = SignalConfig::from_sources(&HashMap::new(), Some(toml)).expect("valid toml");
-        assert_eq!(cfg.endpoint, "10.0.0.5:7583");
-        assert_eq!(cfg.account, "+15551230000");
-        assert_eq!(cfg.allowlist, vec!["+15551230001", "+15551230002"]);
-        assert_eq!(cfg.own_device_id, Some(2));
-        assert!(cfg.reuse_rolling_group);
-        assert_eq!(cfg.rolling_group_id.as_deref(), Some("grp-rolling=="));
-    }
-
-    #[test]
-    fn toml_allowlist_must_be_an_array() {
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = "+15551230001"
-        "#;
-        let err = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap_err();
-        assert!(matches!(err, ConfigError::Toml(_)));
-    }
-
-    #[test]
-    fn toml_allowlist_entries_must_be_strings() {
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001", 123]
-        "#;
-        let err = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap_err();
-        assert!(matches!(err, ConfigError::Toml(_)));
-    }
-
-    #[test]
-    fn env_overrides_toml_per_setting() {
-        let toml = r#"
-            endpoint = "10.0.0.5:7583"
-            account  = "+15550000000"
-            allowlist = ["+15550000001"]
-        "#;
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, Some(toml)).unwrap();
-        // env wins for endpoint + account; allowlist falls back to TOML.
-        assert_eq!(cfg.endpoint, "127.0.0.1:7583");
-        assert_eq!(cfg.account, "+15551230000");
-        assert_eq!(cfg.allowlist, vec!["+15550000001"]);
-    }
-
-    #[test]
-    fn reuse_rolling_group_truthy_env_values() {
-        for v in ["1", "true"] {
-            let e = env(&[
-                (ENV_ENDPOINT, "127.0.0.1:7583"),
-                (ENV_ACCOUNT, "+15551230000"),
-                (ENV_ALLOWLIST, "+15551230001"),
-                (ENV_REUSE_ROLLING_GROUP, v),
-                (ENV_ROLLING_GROUP_ID, "grp-rolling=="),
-            ]);
-            let cfg = SignalConfig::from_sources(&e, None).unwrap();
-            assert!(cfg.reuse_rolling_group, "value {v:?} should be truthy");
-            assert_eq!(cfg.rolling_group_id.as_deref(), Some("grp-rolling=="));
-        }
-    }
-
-    #[test]
-    fn reuse_rolling_group_falsy_env_values_are_per_session() {
-        // Fail-closed: explicit false tokens and empty values must resolve to
-        // per-session isolation, never shared.
-        for v in ["0", "false", "no", "off", "", "  "] {
-            let e = env(&[
-                (ENV_ENDPOINT, "127.0.0.1:7583"),
-                (ENV_ACCOUNT, "+15551230000"),
-                (ENV_ALLOWLIST, "+15551230001"),
-                (ENV_REUSE_ROLLING_GROUP, v),
-            ]);
-            let cfg = SignalConfig::from_sources(&e, None).unwrap();
-            assert!(
-                !cfg.reuse_rolling_group,
-                "value {v:?} must resolve to per-session (reuse=false)"
-            );
-        }
-    }
-
-    #[test]
-    fn unknown_reuse_rolling_group_env_value_is_error() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-            (ENV_REUSE_ROLLING_GROUP, "treu"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(err, ConfigError::Toml(_)));
-    }
-
-    #[test]
-    fn non_boolean_toml_reuse_rolling_group_is_error() {
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001"]
-            reuse_rolling_group = "tru"
-        "#;
-        let err = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap_err();
-        assert!(matches!(err, ConfigError::Toml(_)));
-    }
-
-    #[test]
-    fn absent_reuse_setting_defaults_to_per_session() {
-        // With neither env nor TOML specifying the flag, the default MUST be
-        // per-session (reuse=false) with no bound rolling group id.
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-        ]);
-        let cfg = SignalConfig::from_sources(&e, None).unwrap();
-        assert!(!cfg.reuse_rolling_group);
-        assert_eq!(cfg.rolling_group_id, None);
-
-        // Same via a TOML file that omits the key entirely.
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001"]
-        "#;
-        let cfg = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap();
-        assert!(!cfg.reuse_rolling_group);
-        assert_eq!(cfg.rolling_group_id, None);
-    }
-
-    #[test]
-    fn reuse_rolling_group_opt_in_via_toml() {
-        // Opt-in path: explicit reuse=true + a bound group id in TOML is honored.
-        let toml = r#"
-            endpoint = "127.0.0.1:7583"
-            account  = "+15551230000"
-            allowlist = ["+15551230001"]
-            reuse_rolling_group = true
-            rolling_group_id = "grp-shared=="
-        "#;
-        let cfg = SignalConfig::from_sources(&HashMap::new(), Some(toml)).unwrap();
-        assert!(cfg.reuse_rolling_group);
-        assert_eq!(cfg.rolling_group_id.as_deref(), Some("grp-shared=="));
-    }
-
-    #[test]
-    fn reuse_rolling_group_requires_rolling_group_id() {
-        // Without a pinned group id, "rolling" mode cannot actually roll
-        // across sessions; it would create a fresh group and skip quitGroup.
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-            (ENV_REUSE_ROLLING_GROUP, "1"),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(
-            err,
-            ConfigError::MissingRequired("rolling_group_id")
-        ));
-    }
-
-    #[test]
-    fn empty_rolling_group_id_does_not_enable_reuse() {
-        let e = env(&[
-            (ENV_ENDPOINT, "127.0.0.1:7583"),
-            (ENV_ACCOUNT, "+15551230000"),
-            (ENV_ALLOWLIST, "+15551230001"),
-            (ENV_REUSE_ROLLING_GROUP, "true"),
-            (ENV_ROLLING_GROUP_ID, "  "),
-        ]);
-        let err = SignalConfig::from_sources(&e, None).unwrap_err();
-        assert!(matches!(
-            err,
-            ConfigError::MissingRequired("rolling_group_id")
-        ));
-    }
-
-    #[test]
-    fn malformed_toml_is_error() {
-        let err = SignalConfig::from_sources(&HashMap::new(), Some("this = = broken")).unwrap_err();
-        assert!(matches!(err, ConfigError::Toml(_)));
-    }
-}
+mod tests;
